refactor: add `constants.ts` file
