+++ src/app/(dashboard)/users/_components/UsersTable/TableBody/ReadOnlyRow/index.tsx
@@ -1,5 +1,5 @@
 import { MouseEvent } from "react";
-import { UserProps } from "@/app/(dashboard)/users/_utils/user.type";
+import { UserProps } from "@/app/(dashboard)/users/_utils/_types/user.type";
 import { BsThreeDotsVertical } from "react-icons/bs";
 import { FaTrash } from "react-icons/fa";
