+++ src/app/(dashboard)/users/_utils/_types/user.type.ts
