+++ src/app/(dashboard)/users/_utils/constants.ts
@@ -1,0 +1,2 @@
+export const MAX_VISIBLE_PAGES = 3;
+export const USERS_PER_PAGE = 12;
